+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Update `hidePaginationOnMobile` props to `Carousel`, default to true.
+
 ## [2.59.0] - 2020-03-13
 
 Feature: